% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ki.data.R
\name{as.ki.data}
\alias{as.ki.data}
\title{Convert data set to ki.data object}
\usage{
as.ki.data(input_filepath, start.column = 9L, ...)
}
\arguments{
\item{input_filepath}{Needs documentation.}

\item{start.column}{[\code{integer}] Defaults to \code{9L}.}

\item{...}{Currently not used.}
}
\description{
Convert data set to ki.data object
}
