% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastafrica}
\alias{eastafrica}
\title{East African Sample Data}
\format{
A \code{data.table} with 30684 rows and 6 columns:
\itemize{
\item \code{"PlotId"}: [\code{factor}] Station name (\code{"kil"}: Kilimanjaro,
\code{"jom"}: Nairobi);
\item \code{"Status"}: [\code{factor}] Preprocess level (\code{"cleansed"}: outliers
removed, \code{"filled"}: gaps filled via SSA);
\item \code{"Datetime"}: [\code{POSIXct}] Time in Eastern Africa Time;
\item \code{"TEMP, MIN, MAX"}: [`numeric] Mean, minimum, maximum air temperature in
degC.
}
}
\usage{
eastafrica
}
\description{
A data set containing outlier-adjusted and gap-filled daily readings from the
GSOD stations at Kilimanjaro Intl. Airport, Tanzania, and Jomo Kenyatta
Intl. Airport, Nairobi, Kenya between 1980 and 2000. Missing values were
imputed based on singular spectrum analysis (SSA).
}
\keyword{datasets}
