% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfLinInt.R
\name{gfLinInt}
\alias{gfLinInt}
\title{Imputation of small data gaps by linear interpolation}
\usage{
gfLinInt(data, prm = "TEMP", limit = 5, width = 11, ...)
}
\arguments{
\item{data}{Object of class \code{ki.data}. See \code{\link{as.ki.data}},
\code{\link{gsod2ki}} for further information.}

\item{prm}{Character, default is "TEMP". Parameter(s) to fill.}

\item{limit}{Numeric, default is 5. Maximum gap length to be filled by linear
interpolation.}

\item{width}{Numeric, default is 11. See \code{\link{rollapply}} for further
information.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
An object of class \code{ki.data}.
}
\description{
This function fills small data gaps occurring in (eco-)climatological
measurement series by simple linear interpolation.
}
\examples{
\dontrun{
moshi <- subset(gsodstations, `STATION NAME` == "MOSHI")

gsod_moshi <- dlGsodStations(usaf = moshi$USAF,
                             start_year = 1990, end_year = 1995,
                             dsn = tempdir(),
                             unzip = TRUE)

# Conversion to KiLi SP1 `ki.data` object
ki_moshi <- gsod2ki(data = gsod_moshi,
                    prm_col = c("TEMP", "MIN", "MAX"),
                    df2ki = TRUE)

# Fill small gaps (n <= 3) by linear interpolation
ki_moshi_lf <- gfLinInt(data = ki_moshi,
                        prm = c("TEMP", "MIN", "MAX"),
                        limit = 3, width = 7)

plot(methods::slot(ki_moshi_lf, "Parameter")$TEMP, type = "l", col = "red")
lines(methods::slot(ki_moshi, "Parameter")$TEMP) 
}

}
\seealso{
\code{\link{rollapply}}
}
\author{
Florian Detsch
}
