% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfNogapLength.R
\name{gfNogapLength}
\alias{gfNogapLength}
\title{Julendat function to identify lengths of continuous measurements}
\usage{
gfNogapLength(gaps, data.dep, ...)
}
\arguments{
\item{gaps}{A \code{data.frame} or \code{list}. Usually created from
\code{\link{gfGapLength}}.}

\item{data.dep}{An object of class \code{ki.data}, or a filepath that can be
coerced to \code{ki.data}. The data set under investigation.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
A \code{list} containing start, end, and length of each continuous measurement.
}
\description{
This is a function taken from
\href{https://code.google.com/archive/p/julendat/}{Julendat} to identify
lengths of continuous measurements in (eco-)climatological measurement series.
}
\examples{
\dontrun{
moshi <- subset(gsodstations, `STATION NAME` == "MOSHI")

gsod_moshi <- dlGsodStations(usaf = moshi$USAF,
                             start_year = 1990, end_year = 1995,
                             dsn = tempdir(),
                             unzip = TRUE)

# Conversion to KiLi SP1 `ki.data` object
ki_moshi <- gsod2ki(data = gsod_moshi,
                    prm_col = c("TEMP", "MIN", "MAX"),
                    df2ki = TRUE)

# Identify length per data gap
gaps <- gfGapLength(data.dep = ki_moshi, 
                    pos.na = which(is.na(methods::slot(ki_moshi, "Parameter")$TEMP)), 
                    gap.limit = 365,
                    units = "days", 
                    end.datetime = Sys.Date())

# Identify lengths of continuous measurements
nogaps <- gfNogapLength(gaps = gaps, 
                        data.dep = ki_moshi)
}

}
\seealso{
\code{\link{gfGapLength}}
}
\author{
Florian Detsch
}
