% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfOutputData}
\alias{gfOutputData}
\title{Julendat function to convert \code{ki.data} to \code{data.frame}}
\usage{
gfOutputData(data.dep, plevel = "NA", ...)
}
\arguments{
\item{data.dep}{Object of class \code{ki.data}.}

\item{plevel,}{default is "NA". Processing level. Not required for GSOD data
processing.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A \code{data.frame}.
}
\description{
This function converts objects of class \code{ki.data} to "ordinary" objects
of class \code{data.frame} that follow KiLi SP1 standard formatting conventions.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{gsod2ki}}, \code{\link{as.ki.data}}
}

