% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfRejectLowQuality.R
\name{gfRejectLowQuality}
\alias{gfRejectLowQuality}
\title{Julendat function to assure measurement quality}
\usage{
gfRejectLowQuality(data, prm.dep = "Ta_200", quality.levels, ...)
}
\description{
Julendat function to assure measurement quality
}
