% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfRun.R
\name{gfRun}
\alias{gfRun}
\title{Julendat wrapper function controlling imputation of missing values}
\usage{
gfRun(
  files.dep,
  files.indep,
  filepath.coords,
  quality.levels,
  gap.limit,
  end.datetime,
  units,
  na.limit,
  time.window,
  n.plot,
  prm.dep,
  prm.indep,
  family,
  plevel,
  ...
)
}
\description{
Julendat wrapper function controlling imputation of missing values
}
