% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsodColWidth}
\alias{gsodColWidth}
\title{Fixed column widths of GSOD ASCII data}
\usage{
gsodColWidth(...)
}
\arguments{
\item{...}{Additional arguments. Currently not in use.}
}
\value{
A numeric vector containing fixed column width of GSOD ASCII data.
}
\description{
This function returns the native column widths of GSOD ASCII data, e.g. for
use with \code{\link{read.fwf}}.
}
\examples{
gsodColWidth()
}
\author{
Florian Detsch
}
\seealso{
\code{\link{read.fwf}}
}

