% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsodPlot.R
\name{gsodPlot}
\alias{gsodPlot}
\title{Plot GSOD data}
\usage{
gsodPlot(
  fls_orig = NULL,
  fls,
  stations,
  prm = "TEMP",
  type = c("original", "trends", "both"),
  ...
)
}
\arguments{
\item{fls_orig}{See argument \code{fls}.}

\item{fls}{Character vector containing filepath(s) to the imputed GSOD data
set(s), usually derived from \code{\link{gfLinInt}}, \code{\link{gfJulendat}}
and/or \code{\link{gfSsa}}, or \code{list} containing \code{data.frame} objects
holding the data, or a single \code{data.frame} in case only one GSOD station
shall be visualized.}

\item{stations}{Character. Name(s) of the station(s) corresponding to \code{fls}
and \code{fls_orig} that will be displayed above each facet.}

\item{prm}{Character, default is "TEMP". Determines which parameter to
visualize. If \code{type = "trends"}, this argument will be ignored, and
columns "TEMP", "MIN" and "MAX" must be available in \code{fls} and
\code{fls_orig}.}

\item{type}{Character. Determines the plot type, one of \code{"original"}
(default), \code{"trend"}, or \code{"both"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{element_text}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function aims at visualizing GSOD data, both raw/adjusted for outliers and
imputed, as well as linear trends of mean, minimum and maximum daily air
temperature.
}
\examples{
# Visualize trends in daily mean, minimum, and maximum air temperature
lst = split(
  eastafrica
  , f = paste(
    eastafrica$PlotId
    , eastafrica$Status
    , sep = ","
  )
)

cleansed_data = lst[
  grep(
    "cleansed"
    , names(lst)
  )
]

filled_data = lst[
  grep(
    "filled"
    , names(lst)
  )
]

gsodPlot(
  cleansed_data
  , filled_data
  , stations = c("NAIROBI JKIA", "KILIMANJARO INTL")
  , type = "trends"
)

}
\author{
Florian Detsch
}
