% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsodReformat.R
\name{gsodReformat}
\alias{gsodReformat}
\title{Reformat list of available GSOD stations}
\usage{
gsodReformat(data, df2sp = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame}. Usually built-in \code{gsodstations}.}

\item{df2sp}{Logical, default is `FALSE`. If `TRUE`, the list of available 
GSOD stations is converted to \code{sf}.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
An object of class \code{data.frame}, or a \code{sf} object if 
\code{df2sp = TRUE}.
}
\description{
This function is intended to convert the initial list of available GSOD 
stations (see \url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}) to
an object of class \code{sf} for further processing.
}
\examples{
# Reformat coordinates and convert to spatial object
gsod_shp <- gsodReformat(data = gsodstations, 
                         coords = TRUE, 
                         rm_invalid_coords = TRUE, 
                         df2sp = TRUE)

if (interactive()) {
  plot(gsod_shp)
}

}
\seealso{
\code{\link{gsodDf2Sp}}
}
\author{
Florian Detsch
}
