% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSODTools-package.R
\name{GSODTools-package}
\alias{GSODTools-package}
\title{R Functions to Select, Download, and Process GSOD Data}
\description{
R Functions to Select, Download, and Process GSOD Data
}
\details{
R functions to select, download, and process GSOD data.
}
\author{
Florian Detsch \cr
\cr
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
