% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch2Millimeter.R
\name{inch2Millimeter}
\alias{inch2Millimeter}
\title{Convert inches and hundredths to millimeters}
\usage{
inch2Millimeter(val, ...)
}
\arguments{
\item{val}{Numeric. A vector of precipitation amounts in inches and hundredths.}

\item{...}{Additional arguments passed to \code{\link{round}}.}
}
\value{
A numeric vector of precipitation amounts in mm.
}
\description{
This function converts precipitation amounts from native inches and 
hundredths to millimeters (mm).
}
\examples{
inch2Millimeter(val = c(68, 72, 76), digits = 1)

}
\author{
Florian Detsch
}

