% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outlier2na}
\alias{outlier2na}
\title{Remove statistical outliers from time-series values}
\usage{
outlier2na(val, lower_quantile = 0.2, upper_quantile = 0.8, ...)
}
\arguments{
\item{val}{Numeric. A vector of observed time-series values.}

\item{lower_quantile}{Numeric, default is 0.2. The lower quantile limit.}

\item{upper_quantile}{Numeric, default is 0.8. The upper quantile limit.}

\item{...}{Additional arguments passed to \code{\link{tsOutliers}}.}
}
\value{
An outlier-cleaned numeric vector of time-series values.
}
\description{
This is a wrapper function around \code{\link{tsOutliers}} that automatically
removes identified statistical outliers from a measurement series.
}
\examples{
# Random time-series values
set.seed(10)
x <- rnorm(100, 0, 2)

# Replace outliers with NA
x2 <- outlier2na(x, lower_quantile = .35, upper_quantile = .7)

plot(x, type = "l", lty = 2)
lines(x2, col = "red")
}
\author{
Florian Detsch
}
\seealso{
\code{\link{tsOutliers}}
}

