% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCelsius.R
\name{toCelsius}
\alias{toCelsius}
\title{Convert degree Fahrenheit to degree Celsius}
\usage{
toCelsius(val, ...)
}
\arguments{
\item{val}{Numeric. A vector containing temperature values in degree Fahrenheit.}

\item{...}{Additional arguments passed to \code{\link{round}}.}
}
\value{
A numeric vector containing temperature values in degree Celsius.
}
\description{
This function converts temperature values from native degree Fahrenheit to
degree Celsius.
}
\examples{
toCelsius(val = c(68, 72, 76), digits = 1)

}
\author{
Florian Detsch
}
