% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInventory.R
\name{updateInventory}
\alias{updateInventory}
\title{Retrieve latest version of the GSOD station inventory}
\usage{
updateInventory(online = TRUE)
}
\arguments{
\item{online}{'logical'. If \code{TRUE} (default), the function tries to 
retrieve the latest online version of the GSOD station inventory. Otherwise 
(or if no active internet connection is available), the inventory is imported 
locally.}
}
\value{
A 'data.frame' with information about available GSOD stations.
}
\description{
Download the latest version of the GSOD station inventory from 
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv} or, if no active 
internet connection exists, import the latest local backup.
}
\examples{
updateInventory()

}
\author{
Florian Detsch
}

