% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorHarmonics.R
\name{vectorHarmonics}
\alias{vectorHarmonics}
\title{Fit harmonic trend model to time-series values}
\usage{
vectorHarmonics(
  x,
  frq = 12L,
  st = c(1982, 1),
  nd = c(1986, 12),
  m = 2L,
  fun = stats::median
)
}
\arguments{
\item{x}{Numeric. A vector containing time-series values.}

\item{frq}{Integer, defaults to 12 (months). The number of observations per
unit of time. See \code{\link[stats]{ts}}.}

\item{st}{Numeric. A vector containing start year and month.}

\item{nd}{Numeric. A vector containing end year and month.}

\item{m}{Integer, defaults to 2. The number of pairs of harmonic functions to
be created. See \code{\link[TSA]{harmonic}}.}

\item{fun}{Aggregating function to be applied to the fitted values.}
}
\value{
A numeric vector.
}
\description{
This is a wrapper function around \code{\link[TSA]{harmonic}}
from the \pkg{TSA} package that constructs harmonic functions for fitting a
harmonic trend model.
}
\seealso{
\code{\link{harmonic}}
}
\author{
Tim Appelhans
}
