% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfComputeLinearModel}
\alias{gfComputeLinearModel}
\title{Julendat gap-filling based on linear modelling}
\usage{
gfComputeLinearModel(data = NULL, data.cc = NULL, data.dep,
  family = gaussian, pos.na, plots = NULL, n.plot = 10,
  prm.dep = "Ta_200", prm.indep = NA, time.window.pre, time.window.pre.span,
  time.window.post, time.window.post.span, ...)
}
\description{
#
}

