% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfImputeMissingValue}
\alias{gfImputeMissingValue}
\title{Julendat wrapper function for imputation of missing values}
\usage{
gfImputeMissingValue(data.dep, data.indep, na.limit = 0.2, time.window,
  pos.na, data.coords, n.plot = 10, prm.dep = "Ta_200", prm.indep = NA,
  family = gaussian, ...)
}
\description{
#
}

