% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfValidNaThreshold}
\alias{gfValidNaThreshold}
\title{Julendat function to calculate the amount of missing values}
\usage{
gfValidNaThreshold(data.indep, prm.dep = "Ta_200", na.limit = 0.2,
  time.window.pre, time.window.post, pos.na, ...)
}
\description{
#
}

