% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsod2ki}
\alias{gsod2ki}
\title{Convert GSOD data to Julendat standard format}
\usage{
gsod2ki(data, date_col = "YEARMODA", prm_col = "TEMP", time_step = "day",
  timezone = "NA", aggtime = "-999", plot_id = "NA", ep_plot_id = "xxx",
  station_id = "NA", proc_level = -999, qual_flag = "NA",
  save_output = FALSE, df2ki = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame}. Input data, e.g. from \code{\link{dlGsodStations}}.}

\item{date_col}{Character or numeric, default is "YEARMODA". Indicates the
date column.}

\item{prm_col}{Character or numeric, default is "TEMP". Indicates the
parameter column(s).}

\item{time_step}{Character, default is "day". Indicates the time step between
two measurements.}

\item{timezone}{Character, default is "NA".}

\item{aggtime}{Character, default is "-999".}

\item{plot_id}{Character, default is "NA".}

\item{ep_plot_id}{Character, default is "xxx".}

\item{station_id}{Character, default is "NA".}

\item{proc_level}{Numeric, default is -999.}

\item{qual_flag}{Numeric, default is "NA".}

\item{save_output}{Logical, default is FALSE. If TRUE, a local copy of the
reformatted GSOD data will be created based on the settings provided to
\code{...}}

\item{df2ki}{Logical, default is FALSE. If TRUE, the reformatted data is
converted to an object of class \code{ki.data}.}

\item{...}{Additional arguments passed to \code{\link{write.csv}}.}
}
\value{
An object of class \code{data.frame} or, if \code{df2ki = TRUE}, an object of
class \code{ki.data}.
}
\description{
This function converts GSOD data from native format to standard KiLi SP1 format
(Julendat, see https://code.google.com/p/julendat/) and, optionally, to an
object of class \code{ki.data}.
}
\examples{
library(dplyr)

data(gsodstations)
moshi <- filter(gsodstations, STATION.NAME == "MOSHI")

gsod_moshi <- dlGsodStations(usaf = moshi$USAF,
                             start_year = 1990, end_year = 1995,
                             dsn = paste0(getwd(), "/data/moshi/"),
                             unzip = TRUE)

# Conversion to KiLi SP1 Julendat standard format
jul_moshi <- gsod2ki(data = gsod_moshi,
                     prm_col = c("TEMP", "MIN", "MAX"),
                     df2ki = FALSE)

plot(jul_moshi$TEMP, type = "l", lwd = 2,
     ylim = c(min(jul_moshi$MIN, na.rm = TRUE), max(jul_moshi$MAX, na.rm = TRUE)))
lines(jul_moshi$MIN, col = "blue", lty = 2)
lines(jul_moshi$MAX, col = "red", lty = 2)

# Conversion to `ki.data` object
ki_moshi <- gsod2ki(data = gsod_moshi,
                    prm_col = c("TEMP", "MIN", "MAX"),
                    df2ki = TRUE)

str(ki_moshi)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{as.ki.data}}
}

