% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsodDf2Sp}
\alias{gsodDf2Sp}
\title{Convert GSOD station list to spatial object}
\usage{
gsodDf2Sp(data, ...)
}
\arguments{
\item{data}{\code{data.frame}. Usually \code{data("gsodstations")}.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
An object of class \code{SpatialPointsDataFrame}.
}
\description{
This function converts the list of available GSOD stations from \code{data.frame}
to \code{SpatialPointsDataFrame}.
}
\examples{
# Load data
data(gsodstations)

# Reformat station list
gsod_df <- gsodReformat(gsodstations, df2sp = FALSE)

# Convert station list to spatial object
gsod_shp <- gsodDf2Sp(data = gsod_df)

plot(gsod_shp)
}
\author{
Florian Detsch
}

