% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsodPlot}
\alias{gsodPlot}
\title{Plot GSOD data}
\usage{
gsodPlot(fls_orig = NULL, fls, stations, prm = "TEMP", type = "trends",
  ...)
}
\arguments{
\item{fls_orig}{See argument \code{fls}.}

\item{fls}{Character vector containing filepath(s) to the imputed GSOD data
set(s), usually derived from \code{\link{gfLinInt}}, \code{\link{gfJulendat}}
and/or \code{\link{gfSsa}}, or \code{list} containing \code{data.frame} objects
holding the data, or a single \code{data.frame} in case only one GSOD station
shall be visualized.}

\item{stations}{Character. Name(s) of the station(s) corresponding to \code{fls}
and \code{fls_orig} that will be displayed above each facet.}

\item{prm}{Character, default is "TEMP". Determines which parameter to
visualize. If \code{type = "trends"}, this argument will be ignored, and
columns "TEMP", "MIN" and "MAX" must be available in \code{fls} and
\code{fls_orig}.}

\item{...}{Additional arguments passed to \code{\link{element_text}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function aims at visualizing GSOD data, both raw/adjusted for outliers and
imputed, as well as linear trends of mean, minimum and maximum daily air
temperature.
}
\examples{
# Load outlier-adjusted (*gsod*) and gap-filled (*ssa*) data sets from Nairobi
# and Kilimanjaro Airport, 1980-2000
data("data_nairobi_kilimanjaro")

# Visualize trends in daily mean, minimum, and maximum air temperature
gsodPlot(fls_orig = list(df_gsod_nairobi, df_gsod_kilimanjaro),
         fls = list(df_ssa_nairobi, df_ssa_kilimanjaro),
         stations = c("Jomo Kenyatta Intl.", "Kilimanjaro Intl. Airport"),
         type = "trends")
}
\author{
Florian Detsch
}

