% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsodReformat.R
\name{gsodReformat}
\alias{gsodReformat}
\title{Reformat list of available GSOD stations}
\usage{
gsodReformat(
  data,
  elevation = TRUE,
  coords = TRUE,
  rm_invalid_coords = TRUE,
  df2sp = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame}. Usually \code{data("gsodstations")}.}

\item{elevation}{Logical, default is TRUE. Determines whether elevation is
converted from decimeters to meters.}

\item{coords}{Logical, default is TRUE. Determines whether coordinates are
converted from thousandth of a degree longitude/latitude to whole degrees.}

\item{rm_invalid_coords}{Logical, default is TRUE. Determines whether records
with invalid coordinates, i.e. latitudes > 90 / < -90 degree and longitude > 
180 / < -180 degree, are removed from the data set. Note: if \code{coords = FALSE}, 
this argument will automatically be set to FALSE.}

\item{df2sp}{Logical, default is FALSE. If TRUE, the list of available GSOD
stations is converted from \code{data.frame} to \code{SpatialPointsDataFrame} object.}

\item{...}{Additional arguments. Currently not in use.}
}
\value{
An object of class \code{data.frame}, or an object of class 
\code{SpatialPointsDataFrame} if \code{df2sp = TRUE}.
}
\description{
This function is intended to reformat selected columns, i.e. elevation and
coordinates, from the initial list of available GSOD stations (see 
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}). Optionally, the
corresponding \code{data.frame} can then be converted to an object of class
\code{SpatialPointsDataFrame} for further processing.
}
\examples{
# Load data
data(gsodstations)

# Reformat elevation and coordinates, and convert to spatial object.
gsod_shp <- gsodReformat(data = gsodstations, 
                         elevation = TRUE, 
                         coords = TRUE, 
                         rm_invalid_coords = TRUE, 
                         df2sp = TRUE)
                         
plot(gsod_shp)                        

}
\seealso{
\code{\link{gsodDf2Sp}}
}
\author{
Florian Detsch
}
