% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsodstations}
\alias{gsodstations}
\title{Global Summary of the Day Station List}
\format{
A \code{data.table} with 29640 rows and 11 columns:
\itemize{
\item \code{"USAF"}: [\code{character}] Air Force station ID. May contain a letter in the
first position.
\item \code{"WBAN"}: [\code{character}] NCDC WBAN number.
\item \code{"STATION NAME"}: [\code{character}] Station name.
\item \code{"CTRY"}: [\code{character}] FIPS country ID, see also
\url{https://www.ncei.noaa.gov/pub/data/noaa/country-list.txt}.
\item \code{"STATE"}: [\code{character}] State for US stations.
\item \code{"ICAO"}: [\code{character}] ICAO ID.
\item \code{"LAT, LON"}: [\code{numeric}] Latitude, longitude in thousandths of decimal
degrees.
\item \code{"ELEV(M)"}: [\code{numeric}] Elevation in meters.
\item \code{"BEGIN, END"}: [\code{character}] Beginning, end of Period Of Record (POR;
YYYYMMDD). There may be reporting gaps within the POR.
}
}
\usage{
gsodstations
}
\description{
A data set containing the full Global Summary of the Day (GSOD) station list.
}
\seealso{
Integrated Surface Database Station History, May 2022. Available
online: \url{https://www.ncei.noaa.gov/pub/data/noaa/isd-history.csv} (last
accessed 2023-12-12).
}
\keyword{datasets}
