% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsodstations}
\alias{gsodstations}
\title{Global Summary of the Day Station List}
\format{
A \code{data.table} with 29610 rows and 11 columns.
}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}
}
\usage{
gsodstations
}
\description{
A data set containing the full Global Summary of the Day (GSOD) station list.
}
\keyword{datasets}
