% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSODTools-package.R
\docType{package}
\name{GSODTools-package}
\alias{GSODTools-package}
\alias{GSODToolspackage}
\title{Download and Process Global Surface Summary of the Day Data}
\description{
Download and Process Global Surface Summary of the Day Data
}
\details{
We provide a set of functions to retrieve information about Global Surface 
Summary of the Day (GSOD) weather stations currently available online; 
download and pre-process datasets; and perform various operations in the 
context of time series analysis.
}
\author{
Florian Detsch
}
\references{
NOAA (2016) Global Surface Summary of the Day, 
\url{https://data.noaa.gov/dataset/global-surface-summary-of-the-day-gsod}.
}
\keyword{package}

