% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gzGsodStations}
\alias{gzGsodStations}
\title{Extract and merge downloaded GSOD data}
\usage{
gzGsodStations(usaf, dsn = ".", start_year = NULL, end_year = NULL,
  save_output = FALSE, rm_gz = FALSE, ...)
}
\arguments{
\item{usaf}{Numeric. A unique USAF station code. It can either be manually
determined from \code{data(gsodstations)}, or retreived from spatial
subsetting (see \code{\link{stationFromCoords}},
\code{\link{stationFromExtent}}).}

\item{dsn}{Character, default is the current working directory. Destination
folder for data download.}

\item{start_year}{Numeric. The desired year to start data extraction and
concatenation. If not supplied, all files in \code{dsn} will be considered.}

\item{end_year}{Numeric. The desired year to stop data extraction and
concatenation. If not supplied, all files in \code{dsn} will be considered.}

\item{save_output}{Logical, default is FALSE. If TRUE, a local copy of the
concatenated GSOD data will be created based on the settings provided to \code{...}}

\item{rm_gz}{Logical, default is FALSE. If TRUE, *.gz files are removed after
extraction and concatenation. If \code{unzip = FALSE}, this argument is ignored.}

\item{...}{Additional arguments passed to \code{\link{write.table}}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Extract and merge previously downloaded GSOD data based on a unique USAF
station code and a supplied temporal range.
}
\examples{
library(dplyr)

data(gsodstations)
moshi <- filter(gsodstations, STATION.NAME == "MOSHI")

# Download data from Moshi, Tanzania, from 1990 to 1995
gsod_moshi <- gzGsodStations(usaf = moshi$USAF,
                             start_year = 1990, end_year = 1995,
                             dsn = paste0(getwd(), "/data/moshi/"),
                             save_output = TRUE,
                             file = paste0(getwd(), "/data/moshi/moshi_1990_1995.csv"),
                             row.names = FALSE)

# Plot temperature data (but: time series not continuous!)
plot(gsod_moshi$TEMP, type = "l")
}
\author{
Florian Detsch
}

