% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seasonalShift}
\alias{seasonalShift}
\title{Plot seasonal shifts in GSOD temperature data}
\usage{
seasonalShift(fls, start = c("1973-01-01", "1977-12-31"),
  end = c("2008-01-01", "2012-12-31"), prm = "TEMP", stations, ...)
}
\arguments{
\item{fls}{Character vector containing filepath(s) to the GSOD data set(s) to
visualize, or \code{list} containing \code{data.frame} objects
holding the data, or a single \code{data.frame} in case only one GSOD station
shall be visualized.}

\item{start}{Character. A vector containing the temporal range of the start
interval to fit a harmonic trend model to.}

\item{end}{Character. A vector containing the temporal range of the end
interval to fit a harmonic trend model to.}

\item{prm}{Character, default is "TEMP". Determines which parameter to
visualize.}

\item{stations}{Character. Name(s) of the station(s) corresponding to
\code{fls} that will be displayed above each facet.}

\item{...}{Additional arguments passed to \code{\link{element_text}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function aims at visualizing seasonal shifts in GSOD data aggregated on
a monthly level. The seasonal shifts are derived from a harmonic trend model
fitted to the desired start and end range of a measurement series.
}
\examples{
# Load outlier-adjusted (*gsod*) and gap-filled (*ssa*) data sets from Nairobi
# and Kilimanjaro Airport, 1980-2000
data("data_nairobi_kilimanjaro")

# Visualize seasonal shifts in monthly averaged air temperature
seasonalShift(fls = list(df_gsod_nairobi, df_gsod_kilimanjaro),
              start = c("1980-01-01", "1983-12-31"),
              end = c("1998-01-01", "2000-12-31"),
              stations = c("Jomo Kenyatta Intl.", "Kilimanjaro Intl. Airport"),
              prm = "MAX")
}
\author{
Florian Detsch
}

