% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stationFromCoords}
\alias{stationFromCoords}
\title{Identify GSOD stations in the vicinity of a given geographic location}
\usage{
stationFromCoords(x, y = NULL, width = 50, ...)
}
\arguments{
\item{x}{Numeric. A a single number specifying the x (longitude) coordinate,
or a vector of two numbers specifying the x and y (latitude) coordinate, or
an object of class \code{SpatialPoints}.}

\item{y}{Numeric, default is NULL. A single number specifying the y
coordinate. Note: if a vector of two numbers is supplied to the previous
argument, this parameter will automatically be ignored.}

\item{width}{Numeric, default is 50. The desired buffer width (km) that will
be applied to the specified coordinate.}

\item{...}{Additional arguments passed to \code{\link{round}}.}
}
\value{
An object of class \code{SpatialPointsDataFrame}.
}
\description{
This function allows the selection of a subset of GSOD stations based on a
specified location and a desired buffer size.
}
\examples{
# Identify GSOD stations that lie within a distance of 100 km around Kibo summit
gsod_shp <- stationFromCoords(x = 37.359031,
                              y = -3.065053,
                              width = 100)

mapGriddedData(mapRegion = "africa", plotData = FALSE, borderCol = "black",
               addLegend = FALSE)
points(gsod_shp, col = "red", pch = 20, cex = 2)

# Alternatively
gsod_shp <- stationFromCoords(x = c(37.359031, -3.065053),
                              width = 100)

# Alternatively
kibo <- data.frame(x = 37.359031, y = -3.065053)
coordinates(kibo) <- ~ x + y
projection(kibo) <- CRS("+init=epsg:4326")

gsod_shp <- stationFromCoords(x = kibo,
                              width = 100)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{geodist}}
}

