% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationFromExtent.R
\name{stationFromExtent}
\alias{stationFromExtent}
\title{Identify GSOD stations within spatial extent}
\usage{
stationFromExtent(bb, ...)
}
\arguments{
\item{bb}{Object of class \code{extent}. Usually NULL and constructed from 
the user-drawn bounding box. If supplied, a subset of GSOD stations will be
created without any further user-specified input.}

\item{...}{Currently not used.}
}
\value{
A \code{sf} object.
}
\description{
This function allows the selection of a subset of GSOD stations based on a 
given map extent.
}
\examples{
# Predefined bounding box around Mt. Kilimanjaro region
kili <- sf::st_bbox(c(xmin = 37, xmax = 37.72, ymin = -3.4, ymax = -2.84))
gsod_shp_kili <- stationFromExtent(bb = kili)

}
\seealso{
\code{\link[sf]{st_crop}}
}
\author{
Florian Detsch
}
