% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tsOutliers}
\alias{tsOutliers}
\title{Identification of statistical outliers in time series}
\usage{
tsOutliers(x, lower_quantile = 0.2, upper_quantile = 0.8, plot = FALSE,
  index = FALSE, ...)
}
\arguments{
\item{x}{Numeric. A vector of observed time-series values.}

\item{lower_quantile}{Numeric, default is 0.2. The lower quantile limit.}

\item{upper_quantile}{Numeric, default is 0.8. The upper quantile limit.}

\item{plot}{Logical, default is FALSE. If TRUE, a time-series plot including
identified outliers is generated.}

\item{index}{Logical, default is FALSE. If TRUE, a vector holding the indices
of identified outliers is returned rather than the statistically obtained
scores for each measured value.}

\item{...}{Additional arguments passed to \code{\link{ts}}.}
}
\value{
A numeric vector of scores or, if \code{index = TRUE}, a vector holding the
indices of identified outliers.
}
\description{
This function identifies statistical outliers in a \code{ts} object based on
upper and lower quantile criteria. The function body is mainly taken from
\link{http://stats.stackexchange.com/questions/1142/simple-algorithm-for-online-outlier-detection-of-a-generic-time-series}.
}
\examples{
# Random time-series values
set.seed(10)
x <- rnorm(100, 0, 2)

# Return indices of outliers incl. visualization
tsOutliers(x, lower_quantile = .35, upper_quantile = .7,
           plot = TRUE, index = TRUE)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{ts}}
}

